backend/requirements.txt
django
django-cors-headers
django-filter
djangorestframework
djangorestframework-simplejwt[crypto]
drf-spectacular
gunicorn
psycopg2-binary
uvicorn
